Sorting methods to different small files
